pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "job-search-agent"
version = "0.1.0"
description = "Local, open-source agentic AI job finder with daily email reports"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-ollama>=0.2.0",
    "langgraph>=0.2.0",
    "chromadb>=0.5.0",
    "pydantic>=2.0",
    "feedparser>=6.0",
    "beautifulsoup4>=4.12",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "lxml>=5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]
serve = [
    "langserve[all]>=0.3.0",
    "fastapi>=0.110",
    "uvicorn>=0.29",
]

[project.scripts]
job-search = "main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "main*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
